Move attaching of magazines to subscription types
Until now it was part of subscription type form (create & edit). Since
it has dependency on IssuesModule - I moved it from SubscriptionsModule
to widget in IssuesModule. Widget was already used for displaying
attached magazines on SubscriptionType show screen.

Collateral damage for now is that attaching of Magazines was moved from
edit screen to show screen. But it fits together with subscription type
items so maybe no refactoring will be needed.

remp/crm#332